Initial import
Starting again with a new name - this includes the basic array
functionality and some early notes for Tickit table-specific
features, web pieces have been left out this time.

So far the focus is on the "ordered list" subclass, once the API for
that is stable and working with Tickit/web views we should be able to
bring some of the other data structures in.

git-svn-id: http://entitymodel.com/svn/trunk/Adapter-Async@2579 c5bf8c3b-bf20-4ed3-b413-d5547a49f493